removed test route in user_routes. renamed app directory to backend.
