+++ backend/__init__.py
@@ -5,14 +5,14 @@ from flask_migrate import Migrate
 from flask_wtf.csrf import CSRFProtect, generate_csrf
 from flask_login import LoginManager
 
-from .models import db, User
-from .api.user_routes import user_routes
-from .api.auth_routes import auth_routes
-from .socket import socketio
+from backend.models import db, User
+from backend.api.user_routes import user_routes
+from backend.api.auth_routes import auth_routes
+from backend.socket import socketio
 
-from .seeds import seed_commands
+from backend.seeds import seed_commands
 
-from .config import Config
+from backend.config import Config
 
 app = Flask(__name__)
