+++ backend/models/book_club_books.py
